Refactor : User, Article 패키지 변경
User 와 Article domain 패키지로 변경
