+++ src/main/java/kr/codesqaud/cafe/controller/UserController.java
@@ -1,7 +1,7 @@
 package kr.codesqaud.cafe.controller;
 
 import kr.codesqaud.cafe.repository.SignUpService;
-import kr.codesqaud.cafe.user.User;
+import kr.codesqaud.cafe.domain.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +11,6 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.PostConstruct;
 import java.util.List;
-import java.util.Optional;
 
 @Controller
 @RequestMapping("/users")
