+++ src/main/java/kr/codesqaud/cafe/domain/User.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe.user;
+package kr.codesqaud.cafe.domain;
 
 public class User {
     private String userId;
