+++ src/main/java/kr/codesqaud/cafe/repository/ArticleRepository.java
@@ -1,7 +1,6 @@
 package kr.codesqaud.cafe.repository;
 
-import kr.codesqaud.cafe.article.Article;
-import org.springframework.stereotype.Repository;
+import kr.codesqaud.cafe.domain.Article;
 
 import java.util.List;
 import java.util.Optional;
