+++ src/main/java/kr/codesqaud/cafe/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package kr.codesqaud.cafe.repository;
 
-import kr.codesqaud.cafe.user.User;
+import kr.codesqaud.cafe.domain.User;
 
 import java.util.List;
 import java.util.Optional;
