+++ src/main/java/kr/codesqaud/cafe/repository/SignUpService.java
@@ -1,6 +1,6 @@
 package kr.codesqaud.cafe.repository;
 
-import kr.codesqaud.cafe.user.User;
+import kr.codesqaud.cafe.domain.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
