+++ src/test/java/kr/codesqaud/cafe/domain/repository/UserSaveTest.java
@@ -2,7 +2,7 @@ package kr.codesqaud.cafe.domain.repository;
 
 import kr.codesqaud.cafe.repository.MemoryMemberRepository;
 import kr.codesqaud.cafe.repository.SignUpService;
-import kr.codesqaud.cafe.user.User;
+import kr.codesqaud.cafe.domain.User;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
